📐moved componets off app.componet
