+++ Frontend/src/app/app.component.html
@@ -8,7 +8,7 @@
     <app-header></app-header>
   </div>
   <div class="header-container col-start-5 col-span-1 row-start-1 row-span-1">
-    <app-header></app-header>
+    <app-other-nav></app-other-nav>
   </div>
         <app-navbar class="col-start-1 col-span-1 row-start-1 pl-16"></app-navbar>
         <app-side-bar class="col-start-1 col-span-1 row-start-2 row-span-4" *ngIf="screenSize === 'desktop'"></app-side-bar>
