+++ Frontend/src/app/components/templates/desktop/other-nav/other-nav.component.html
@@ -1,1 +1,5 @@
-<p>other-nav works!</p>
+<div class="flex items-center">
+  <app-info-icon></app-info-icon>
+  <app-mood-drop-down></app-mood-drop-down>
+  <app-profile-atomic></app-profile-atomic>
+</div>
