+++ Frontend/src/app/shared/setting-pages/display/display.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ThemeService } from '../../../services/theme.service';
 import { NgIf, NgClass } from '@angular/common';
-import { MoodDropDownComponent } from '../../mood-drop-down/mood-drop-down.component';
+import { MoodDropDownComponent } from '../../../components/organisms/mood-drop-down/mood-drop-down.component';
 import { MoodService } from '../../../services/mood-service.service';
 
 @Component({
